Parse range as a sequence of start+dots+end
This was not trivial. The unquoted string symbol had to be reworked, and most
argument types now need a hidden version.

Also had to disable the yang-version type, since it now conflicts with integer
type. I will fix this by making a separate "yang-version" statement.

See this discussion for context:

    https://github.com/tree-sitter/tree-sitter/discussions/1287